Added Vue.js sample site that connects to Auth0
This is just the boilerplate Vue.js sample to connect to Auth0; assuming
that Vue.js is going to be one of our front ends.

I believe that Auth0 needs to setup callback URLs in the App Settings.
The following dev settings:

Callback URL: http://localhost:3000/callback
Allowed Origins: http://localhost:3000
Allowed Logout: http://localhost:3000

Then I _think_ it should work.

Boot this app with:

npm install
npm run serve

Then you can navigate to http://localhost:3000

The Login button is not visible yet, but it should be when the above
settings are changed in Auth0.