Remove filters and upgrade MassTransit
I've removed the security filters from the top of each method as they
are a bit magic.  They have been replaced with a combination of
custom model binding and extension methods.  Personally I think
it's much easier to follow.

This commit also changes MassTransit to use the new mediator stuff.